feat - Add admin rights and 2 charts

+++ back/app.py
@@ -6,6 +6,7 @@ from flask_cors import CORS
 from kaga_logger import DEBUG, Logger
 from peewee import DoesNotExist, IntegrityError
 from werkzeug.exceptions import HTTPException
+from dateutil.rrule import rrule, MONTHLY
 
 from conf import PAGE_SIZE
 from utils import token_protected, check_token, check_admin
@@ -44,7 +45,8 @@ def token():
 @APP.route('/users', methods=['POST', 'GET'])
 def user_list():
     if request.method == "GET":
-        check_token()
+        if not check_admin():
+            abort(403, "Admin rights are required")
         page = int(request.args.get('page') or 0)
         users = User.select().order_by(User.username).offset(page * PAGE_SIZE).limit(PAGE_SIZE)
         return jsonify([user.json() for user in users])
@@ -66,7 +68,8 @@ def user_detail(username: str):
         except DoesNotExist:
             abort(404, f"User '{username}' not found")
     if request.method == "DELETE":
-        check_admin()
+        if not check_admin():
+            abort(403, "Admin rights are required")
         try:
             user = User.get(User.username == username)
             user.delete_instance()
@@ -98,12 +101,40 @@ def record_list():
 @token_protected
 def record_detail(record_id: int):
     if request.method == 'DELETE':
+        if not check_admin():
+            abort(403, "Admin rights are required")
         try:
             record = Record.get_by_id(record_id)
             record.delete_instance()
             return jsonify(), 204
         except DoesNotExist:
             abort(404, f"Record '{record_id}' not found'")
+
+@APP.route('/charts')
+@token_protected
+def charts():
+    if request.method == 'GET':
+        # column: number per month
+        start_date = Record.select().order_by(Record.date).get().date#.strftime('%Y-%m')
+        end_date = Record.select().order_by(-Record.date).get().date#.strftime('%Y-%m')
+        column_chart = [['Month', 'Number of records per month']]
+        for date in rrule(freq=MONTHLY, dtstart=start_date, until=end_date):
+            column_chart.append([
+                date.strftime('%Y-%m'),
+                Record.select().where(Record.date.between(
+                    date.strftime('%Y-%m') + '-01',
+                    date.strftime('%Y-%m') + '-31')
+                ).count()
+            ])
+        # histogram
+        histogram_chart = [['Username', 'Best record']]
+        for user in User.select():
+            if user.best_record:
+                histogram_chart.append([user.username, user.best_record.chrono])
+        return jsonify(
+            columnChart=column_chart,
+            histogramChart=histogram_chart
+        )
 
 if __name__ == "__main__":
     try:
